fleetctrl/backend/requirements.txt
fastapi
uvicorn[standard]
pyyaml
paramiko
docker
pydantic
websockets
python-multipart
